Add method for multiple visual and collision props
The attributes 'visual' and 'collision' exist only for back-compability reasons
and shall only be used if there exists one and only one visual or collision element.
For links with multiple visual or collision elements the new method is added which
shall be used. This simplifies the usage of the API.